% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVfunction.R
\name{CVfunction}
\alias{CVfunction}
\title{For the gap between the predicted value and expected value of the model, the model validates the function}
\usage{
CVfunction(pred,actual)
}
\arguments{
\item{pred}{: Value predicted by the model}
\item{actual}{: The real value}
}
\description{
For the gap between the predicted value and expected value of the model, the model validates the function
}
\examples{
test.pred <- c(2,4,5,7,2,4)
test.obs <- c(1,2,3,4,5,6)
myres <- CVfunction(test.pred,test.obs)
print(myres)
}
