% Please edit documentation in R/data.R
\docType{data}
\name{df.dem}
\alias{df.dem}
\title{Sampling test data of the dem}
\format{
A data frame with 211415 rows and 3 variables:
\describe{
\item{dem}{data variable DEM}
\item{x}{The coordinate variable x}
\item{y}{The coordinate variable y}
}
}
\usage{
df.dem
}
\description{
A dataset containing the df.dem and other attributes of almost 212000
df.dem The variables are as follows:
}
\keyword{datasets}
