% Please edit documentation in R/data.R
\docType{data}
\name{df.input}
\alias{df.input}
\title{Sampling test data}
\format{
A data frame with 109 rows and 6 variables:
\describe{
\item{socd030}{data variable socd030}
\item{dem}{data variable DEM}
\item{plancur}{data variable plancur}
\item{procur}{data variable procur}
\item{mrrtf}{data variable mrrtf}
\item{twi}{data variable twi}
}
}
\usage{
df.input
}
\description{
A dataset containing the testdata and other attributes of almost 110
socd030 The variables are as follows:
}
\keyword{datasets}
