% Please edit documentation in R/data.R
\docType{data}
\name{df.mrrtf}
\alias{df.mrrtf}
\title{Sampling test data of the mrrtf}
\format{
A data frame with 211415 rows and 3 variables:
\describe{
\item{mrrtf}{data variable MRRTF}
\item{x}{The coordinate variable x}
\item{y}{The coordinate variable y}
}
}
\usage{
df.mrrtf
}
\description{
A dataset containing the df.mrrtf and other attributes of almost 212000
df.mrrtf The variables are as follows:
}
\keyword{datasets}
