% Please edit documentation in R/data.R
\docType{data}
\name{df.plancur}
\alias{df.plancur}
\title{Sampling test data of the plancur}
\format{
A data frame with 211415 rows and 3 variables:
\describe{
\item{plancur}{data variable PLANCUR}
\item{x}{The coordinate variable x}
\item{y}{The coordinate variable y}
}
}
\usage{
df.plancur
}
\description{
A dataset containing the df.plancur and other attributes of almost 212000
df.plancur The variables are as follows:
}
\keyword{datasets}
