% Please edit documentation in R/data.R
\docType{data}
\name{df.procur}
\alias{df.procur}
\title{Sampling test data of the procur}
\format{
A data frame with 211415 rows and 3 variables:
\describe{
\item{procur}{data variable PROCUR}
\item{x}{The coordinate variable x}
\item{y}{The coordinate variable y}
}
}
\usage{
df.procur
}
\description{
A dataset containing the df.procur and other attributes of almost 212000
df.procur The variables are as follows:
}
\keyword{datasets}
