% Please edit documentation in R/data.R
\docType{data}
\name{df.twi}
\alias{df.twi}
\title{Sampling test data of the twi}
\format{
A data frame with 211415 rows and 3 variables:
\describe{
\item{twi}{data variable TWI}
\item{x}{The coordinate variable x}
\item{y}{The coordinate variable y}
}
}
\usage{
df.twi
}
\description{
A dataset containing the df.twi and other attributes of almost 212000
df.twi The variables are as follows:
}
\keyword{datasets}
