% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPredictorSubset.R
\name{GetPredictorSubset}
\alias{GetPredictorSubset}
\title{calculation function for cutting spatial data (tool function,Not as an open function, only for function calls)}
\usage{
GetPredictorSubset(
  predictor.name,
  iblock,
  nblock,
  fn,
  nr,
  nc,
  resolutions,
  pro,
  from,
  to
)
}
\arguments{
\item{predictor.name}{: the name of the predictor variable}

\item{iblock}{: sequence code of parallel computing}

\item{nblock}{: number of target blocks (integer)}

\item{fn}{: The passed value of a global variable}

\item{nr}{: The passed value of a global variable}

\item{nc}{: The passed value of a global variable}

\item{resolutions}{: The passed value of a global variable}

\item{pro}{: The passed value of a global variable}

\item{from}{: Which row to start cutting the matrix}

\item{to}{: Where does the last row of the cut matrix go}
}
\value{
Parallel calculation of the cut part of the data box data
}
\description{
calculation function for cutting spatial data (tool function,Not as an open function, only for function calls)
}
\examples{
\donttest{
GetPredictorSubset("dem",4,10,"covariate",486,777,NULL,NULL,1,10)
}
}
\references{
Breiman, L. (2001). Random forests. Mach. Learn. 45, 5–32.
Meinshausen, N. (2006) "Quantile Regression Forests", Journal of Machine Learning Research 7,
983-999 http://jmlr.csail.mit.edu/papers/v7/
Song, X.D., Ge, G.Q., Zhang, G.L. and Wu, H.Y. ParallelDSM: A R package for parallel soil mapping. Computers & Geosciences (to be available in 2021)
}
