% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsepectionVariable.R
\name{Insepect_MLR}
\alias{Insepect_MLR}
\title{A function that checks the parallel computation for missing data of MLR model.}
\usage{
Insepect_MLR(block, outputDirectory)
}
\arguments{
\item{block}{: The number of blocks for data cutting.}

\item{outputDirectory}{: The directory of output files.}
}
\description{
A function that checks the parallel computation for missing data of MLR model.
}
\examples{
\donttest{
Insepect_MLR(30, "./MlrOutput")
}

}
