% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsepectionVariable.R
\name{Insepect_QRF}
\alias{Insepect_QRF}
\title{A function that checks the parallel computation for missing data of QRF model.}
\usage{
Insepect_QRF(block, outputDirectory)
}
\arguments{
\item{block}{: The number of blocks for data cutting.}

\item{outputDirectory}{: The directory of output files.}
}
\description{
A function that checks the parallel computation for missing data of QRF model.
}
\examples{
\donttest{
Insepect_QRF(30, "./QrfOutput")
}

}
