% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsepectionVariable.R
\name{Insepect_RF}
\alias{Insepect_RF}
\title{A function that checks the parallel computation for missing data of RF model.}
\usage{
Insepect_RF(block, outputDirectory)
}
\arguments{
\item{block}{: The number of blocks for data cutting.}

\item{outputDirectory}{: The directory of output files.}
}
\description{
A function that checks the parallel computation for missing data of RF model.
}
\examples{
\donttest{
Insepect_RF(30, "./RfOutput")
}

}
