% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsepectionVariable.R
\name{InsepectionVariable}
\alias{InsepectionVariable}
\title{A function that checks the parallel computation for missing data}
\usage{
InsepectionVariable(model = "MLR", block, outputDirectory)
}
\arguments{
\item{model}{: The models were selected, including QRF,RF and MLR.}

\item{block}{: The number of blocks for data cutting.}

\item{outputDirectory}{: The directory of output files.}
}
\description{
A function that checks the parallel computation for missing data
}
\examples{
\donttest{
InsepectionVariable(model = "MLR", block = 30, outputDirectory = "MlrOutput")
}



}
