% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelInit.R
\name{NormalizeData}
\alias{NormalizeData}
\title{Standardize and normalize data elements}
\usage{
NormalizeData()
}
\description{
Standardize and normalize data elements
}
\examples{
\donttest{
# This function is optional to the user, depending on the data situation.
NormalizeData()
}


}
