% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{DataProcess}
\alias{DataProcess}
\title{Parallel computing initialization preparation(This function is not open to users!)}
\usage{
DataProcess(mymodel)
}
\arguments{
\item{mymodel}{: The models were selected, including QRF,RF and MLR.}
}
\value{
Represents whether the loading of the required variables and dependent packages is complete
}
\description{
Parallel computing initialization preparation(This function is not open to users!)
}
\examples{
\donttest{
#This function only serves the dsmParallel function.
DataProcess(mymodel = "QRF")
}
}
