% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{dsmParallel}
\alias{dsmParallel}
\title{dsmparallel computings}
\usage{
dsmParallel(outpath)
}
\arguments{
\item{outpath}{: Output path of the result of the prediction file. The default is "output".}
}
\description{
dsmparallel computings
}
\examples{
\donttest{
dsmParallel(outpath = "myoutputs")
}
}
\references{
Breiman, L. (2001). Random forests. Mach. Learn. 45, 5–32.
Meinshausen, N. (2006) "Quantile Regression Forests", Journal of Machine Learning Research 7,
983-999 http://jmlr.csail.mit.edu/papers/v7/
Song, X.D., Ge, G.Q., Zhang, G.L. and Wu, H.Y. ParallelDSM: A R package for parallel soil mapping. Computers & Geosciences (to be available in 2021)
}
