% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsepectionVariable.R
\name{InsepectionVariable}
\alias{InsepectionVariable}
\title{A function that checks the parallel computation for missing data}
\usage{
InsepectionVariable(myblock)
}
\arguments{
\item{myblock}{: the number of blocks for data cutting}
}
\description{
A function that checks the parallel computation for missing data
}
\examples{
\donttest{
InsepectionVariable(myblock = 10)
}



}
