% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{Parallelinit}
\alias{Parallelinit}
\title{Parallel computing initialization preparation}
\usage{
Parallelinit(mymodel)
}
\arguments{
\item{mymodel}{: The models were selected, including QRF,RF and MLR.}
}
\value{
Represents whether the loading of the required variables and dependent packages is complete
}
\description{
Parallel computing initialization preparation
}
\examples{
\donttest{
Parallelinit(mymodel = "QRF")
}
}
