% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{Parallelinit}
\alias{Parallelinit}
\title{Parallel computing initialization preparation}
\usage{
Parallelinit()
}
\value{
Represents whether the loading of the required variables and dependent packages is complete
}
\description{
Parallel computing initialization preparation
}
\examples{
\donttest{
Parallelinit()
}
}
