% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smalltesttoy.R
\name{smalltesttoy}
\alias{smalltesttoy}
\title{Black box test function to test whether R package was installed successfully}
\usage{
smalltesttoy(myflag)
}
\arguments{
\item{myflag}{: The black box tests the successful entry mark}
}
\description{
Black box test function to test whether R package was installed successfully
}
\examples{
smalltesttoy(myflag = "1")
}
